perf(fonts): improve performance by preloading fonts and avoid FOIT
The `<link rel="preload">` makes fonts to load faster and reduce layout shifts. the crossorigin attribute is needed, [see](https://stackoverflow.com/a/70878420/2149553)

About the `font-display: swap` property in `@font-face`:

- When the page loads, the browser immediately displays fallback text using a system font instead of hiding the text.

- As soon as the custom font is downloaded, the browser "swaps" the fallback font with the custom font.

If you don’t set font-display, many browsers will hide text (invisible) for up to 3 seconds (this is called `FOIT` – Flash of Invisible Text), waiting for the font to load. If it fails to load in time, fallback fonts are used.